+++ examples/architectureComponentsExample/src/main/java/io/realm/examples/arch/PersonListFragment.java
@@ -114,7 +114,7 @@ public class PersonListFragment extends Fragment {
                     return;
                 }
                 AppCompatActivity activity = ContextUtils.findActivity(view.getContext());
-                PersonFragment personFragment = PersonFragment.create(person.getName());
+                PersonFragment personFragment = PersonFragment.create(person.name);
                 activity.getSupportFragmentManager()
                         .beginTransaction()
                         .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE)