+++ realm-transformer/src/main/java/io/realm/gradle/RealmPluginExtension.java
@@ -22,13 +22,8 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 
 public class RealmPluginExtension {
-
-    public static final String KEY_SYNC_ENABLED = "syncEnabled";
-    public static final String KEY_KOTLIN_EXTENSIONS_ENABLED = "kotlinExtensionsEnabled";
-
     private boolean syncEnabled;
     private boolean kotlinExtensionsEnabled;
-    private Map<String, PropertyChangedListener> listeners = new LinkedHashMap<>();
 
     @Input
     public boolean isSyncEnabled() {
@@ -37,7 +32,6 @@ public class RealmPluginExtension {
 
     public void setSyncEnabled(boolean syncEnabled) {
         this.syncEnabled = syncEnabled;
-        notifyChange(KEY_SYNC_ENABLED, syncEnabled);
     }
 
     @Input
@@ -47,25 +41,5 @@ public class RealmPluginExtension {
 
     public void setKotlinExtensionsEnabled(boolean kotlinExtensionsEnabled) {
         this.kotlinExtensionsEnabled = kotlinExtensionsEnabled;
-        notifyChange(KEY_KOTLIN_EXTENSIONS_ENABLED, kotlinExtensionsEnabled);
-    }
-
-    public void addPropertyListener(String property, PropertyChangedListener listener) {
-        listeners.put(property, listener);
-    }
-
-    private void notifyChange(String key, Object value) {
-        PropertyChangedListener listener = listeners.get(key);
-        if (listener != null) {
-            // Up to users of the API to use the correct generic type, otherwise it will crash
-            // at runtime.
-            //noinspection unchecked
-            listener.onChange(value);
-        }
-    }
-
-    // Callback triggered when the extension property is changed
-    public interface PropertyChangedListener<T> {
-        void onChange(T value);
     }
 }