+++ realm/realm-library/src/main/cpp/observable_collection_wrapper.hpp
@@ -86,20 +86,10 @@ void ObservableCollectionWrapper<T>::start_listening(JNIEnv* env, jobject j_coll
         m_collection_weak_ref = jni_util::JavaGlobalWeakRef(env, j_collection_object);
     }
 
-    auto cb = [=](CollectionChangeSet const& changes, std::exception_ptr err) {
+    auto cb = [=](CollectionChangeSet const& changes) {
         // OS will call all notifiers' callback in one run, so check the Java exception first!!
         if (env->ExceptionCheck())
             return;
-
-        if (err) {
-            try {
-                std::rethrow_exception(err);
-            }
-            catch (const std::exception& e) {
-                realm::jni_util::Log::e("Caught exception in collection change callback %1", e.what());
-                return;
-            }
-        }
 
         m_collection_weak_ref.call_with_local_ref(env, [&](JNIEnv* local_env, jobject collection_obj) {
             local_env->CallVoidMethod(