+++ realm/realm-library/src/main/java/io/realm/mongodb/sync/SubscriptionSet.java
@@ -36,6 +36,7 @@ import io.realm.internal.objectstore.OsSubscriptionSet;
  */
 @Beta
 @Keep
+@ObjectServer
 public interface SubscriptionSet extends Iterable<Subscription> {
 
     /**