+++ realm/realm-library/src/testUtils/java/io/realm/entities/KeywordFieldNames.java
@@ -1,0 +1,10 @@
+package io.realm.entities;
+
+import io.realm.RealmObject;
+
+public class KeywordFieldNames extends RealmObject {
+    public String desc;
+    public String sort;
+    public String distinct;
+    public String limit;
+}