+++ tools/sync_test_server/app_template/functions/canWritePartition/source.js
@@ -1,0 +1,12 @@
+/**
+ * Users with an email that contains `_nowrite_` do not have write access,
+ * all others do.
+ */
+exports = async (partition) => {
+  const email = context.user.data.email;
+  if (email != undefined) {
+    return(!email.includes("_nowrite_"));
+  } else {
+    return true;
+  }
+}